+++ lista.h
@@ -2,11 +2,11 @@
 #define LISTA_H
 
 typedef struct Lista{
-    int n;
+    unsigned long long n;
     struct Lista * prox;
 }item;
-item * aloca_no(int n);
-void adiciona_no(item * raiz, int n);
+item * aloca_no(unsigned long long n);
+void adiciona_no(item * raiz, unsigned long long n);
 item * desaloca_lista(item * raiz);
 
 #endif /*LISTA_H*/
